Final Work Part 4 (#7)
* Add: creation of record label and updating table
Arvin was not present for any of this...jk

* Add: continuing work on gui

+++ create_tables.sql
@@ -3,40 +3,40 @@ CREATE TABLE recordLabel(
     companyName text,
     dateEstablished date,
     labelLocation text,
-    recordLabelId smallint not null,
+    recordLabelId bigint not null,
     PRIMARY KEY (recordLabelId)
 );
 -- 2. Publishes
 CREATE TABLE publishes(
-    albumId smallint not null,
-    recordLabelId smallint not null,
+    albumId bigint not null,
+    recordLabelId bigint not null,
     PRIMARY KEY (albumId, recordLabelId)
 );
 --3. artist
 CREATE TABLE artist(
-    artistId smallint not null,
+    artistId bigint not null,
     artistName text,
     age integer,
     PRIMARY KEY (artistId)
 );
 --4. musician
 CREATE TABLE musician(
-    artistId smallint not null,
-    musicianId smallint not null,
+    artistId bigint not null,
+    musicianId bigint not null,
     instrument text,
     band text,
     PRIMARY KEY (artistId, musicianId)
 );
 --5. Played
 CREATE TABLE played(
-    albumId smallint not null,
-    musicianId smallint not null,
+    albumId bigint not null,
+    musicianId bigint not null,
     PRIMARY KEY (albumId, musicianId)
 );
 --6. album
 CREATE TABLE album(
     albumDuration integer,
-    albumId smallint not null,
+    albumId bigint not null,
     title text,
     coverArtURL text,
     PRIMARY KEY (albumId)
@@ -46,22 +46,22 @@ CREATE TABLE song(
     title text,
     genre text,
     songDuration integer,
-    songId smallint not null,
+    songId bigint not null,
     sourceLink text,
     releaseYear integer,
     PRIMARY KEY (songId)
 );
 --8. Contains
 CREATE TABLE contains(
-    albumId smallint not null,
-    songId smallint not null,
+    albumId bigint not null,
+    songId bigint not null,
     PRIMARY KEY (songId, albumId)
 );
 --9. made
 CREATE TABLE made(
     knownFor text,
-    albumId smallint not null,
-    artistId smallint not null,
+    albumId bigint not null,
+    artistId bigint not null,
     PRIMARY KEY (albumId, artistId),
     FOREIGN KEY (albumId) REFERENCES album(albumId),
     FOREIGN KEY (artistId) REFERENCES artist(artistId)
@@ -71,8 +71,8 @@ CREATE TABLE rating(
     numOfRating integer,
     averageRating decimal,
     userRating integer,
-    albumId smallint not null,
-    songId smallint not null,
+    albumId bigint not null,
+    songId bigint not null,
     PRIMARY KEY (albumId, songId),
     FOREIGN KEY (songId) REFERENCES song(songId),
     FOREIGN KEY (albumId) REFERENCES album(albumId)
